+++ R/predict_boots.R
@@ -17,12 +17,12 @@
 #'
 #' @export
 #'
+#' @importFrom rlang warn
 #' @importFrom rsample bootstraps
+#' @importFrom purrr map_dfc
 #' @importFrom tibble rowid_to_column
 #' @importFrom tidyr pivot_longer
 #' @importFrom tidyr nest
-#' @importFrom purrr map_dfc
-#' @importFrom rlang :=
 #'
 #' @examples
 #' \dontrun{
@@ -40,16 +40,26 @@
 #'   predict_boots(n = 125, training_data = mtcars, new_data = mtcars)
 #' }
 predict_boots <- function(workflow,
-                          n = 100,
+                          n = 2000,
                           training_data,
                           new_data,
+                          verbose = FALSE,
                           ...) {
 
   # check arguments
   assert_workflow(workflow)
   assert_n(n)
-  assert_pred_data(workflow, training_data)
-  assert_pred_data(workflow, new_data)
+  assert_pred_data(workflow, training_data, "training")
+  assert_pred_data(workflow, new_data, "new")
+
+  # warn if low n
+  if (n < 2000) {
+
+    rlang::warn(
+      paste0("At least 2000 resamples recommended for stable results.")
+    )
+
+  }
 
   # create resamples from training set
   training_boots <-
@@ -68,6 +78,7 @@ predict_boots <- function(workflow,
         workflow = workflow,
         boot_splits = training_boots,
         new_data = new_data,
+        verbose = verbose,
         index = .x
       )
     )
@@ -93,120 +104,37 @@ predict_boots <- function(workflow,
 
 }
 
-#' Append a tibble of predictions returned by `predict_boots()` with a prediction
-#' interval.
-#'
-#' @details Generates a summary of predictions with a upper and lower interval
-#'  range. Presently, the `quantile()` function from the `{stats}` package is
-#'  used to determine the lower, 50th percentile, and upper interval ranges.
-#'
-#' @return Appends the tibble of predictions returned by `predict_boots()` with
-#'  three new columns: `.pred_lower`, `.pred`, and `.pred_upper`.
-#'
-#' @param .data a tibble of predictions returned by `predict_boots()`.
-#' @param conf an integer between (0, 1) specifying the interval range.
-#' @param summary_type currently fixed as "quantile". More options may be added
-#'  in later releases.
-#'
-#' @export
-#'
-#' @importFrom dplyr mutate
-#' @importFrom purrr map
-#' @importFrom stats quantile
-#' @importFrom tidyr unnest
-#' @importFrom tidyr pivot_wider
-#'
-#' @examples
-#' \dontrun{
-#' library(tidymodels)
-#'
-#' # setup a workflow without fitting
-#' wf <-
-#'   workflow() %>%
-#'   add_recipe(recipe(qsec ~ wt, data = mtcars)) %>%
-#'   add_model(linear_reg())
-#'
-#' # fit and predict 125 bootstrap resampled models to mtcars
-#' set.seed(123)
-#' preds <-
-#'   wf %>%
-#'   predict_boots(n = 125, training_data = mtcars, new_data = mtcars)
-#'
-#' # append with prediction interval summary columns
-#' preds %>%
-#'   summarise_predictions(conf = 0.95)
-#' }
-summarise_predictions <- function(.data,
-                                  conf = 0.95,
-                                  summary_type = "quantile") {
-
-  # check arguments
-  assert_summary_data(.data)
-  assert_conf(conf)
-
-  # internal renaming
-  pred_summary <- .data
-
-  # determine ci_lower & ci_upper values from conf
-  ci_lower <- (1 - conf)/2
-  ci_upper <- ci_lower + conf
-
-  # return max row
-  n_rows <- nrow(pred_summary)
-
-  # add interval - returns nested col
-  if (summary_type == "quantile") {
-
-    pred_summary <-
-      dplyr::mutate(
-        pred_summary,
-        interval = purrr::map(pred_summary$.preds, ~stats::quantile(.x$model.pred, probs = c(ci_lower, 0.5, ci_upper)))
-      )
-
-  } else {
-
-    message("Other summary_types will be added as part of new releases.")
-    return()
-
-  }
-
-  # add interval labels & display as their own cols
-  pred_summary <- tidyr::unnest(pred_summary, interval)
-  pred_summary <-
-    dplyr::mutate(
-      pred_summary,
-      pred_level = rep(c(".pred_lower", ".pred", ".pred_upper"), n_rows)
-    )
-
-  pred_summary <-
-    tidyr::pivot_wider(
-      pred_summary,
-      names_from = pred_level,
-      values_from = interval
-    )
-
-  return(pred_summary)
-
-}
-
 # --------------------------------internals-------------------------------------
 
-#' Fit a model and predict based on a single bootstrap resample
+#' (Internal) Generate a column of predictions on new data based on a model fit
+#' to a single training bootstrap.
 #'
-#' @param workflow An un-fitted workflow object.
-#' @param boot_splits A bootstrap split object created by `rsample::bootstraps()`.
-#' @param new_data New data to make predictions
-#' @param index Index of `boot_splits` to use for training
+#' @param workflow passed from `predict_boots()`
+#' @param boot_splits passed from `predict_boots()`
+#' @param new_data passed from `predict_boots()`
+#' @param verbose passed from `predict_boots()`
+#' @param index passed from `predict_boots()`
 #'
 #' @importFrom rsample training
+#' @importFrom rsample testing
 #' @importFrom generics fit
+#' @importFrom dplyr filter
+#' @importFrom dplyr pull
 #' @importFrom stats predict
-#' @importFrom dplyr rename
 #' @importFrom rlang sym
+#' @importFrom tidyr crossing
+#' @importFrom Metrics rmse
+#' @importFrom stats sd
+#' @importFrom tibble add_column
+#' @importFrom stats rnorm
+#' @importFrom dplyr mutate
+#' @importFrom dplyr rename
+#' @importFrom rlang :=
 #'
 predict_single_boot <- function(workflow,
                                 boot_splits,
                                 new_data,
+                                verbose,
                                 index) {
 
   # get training data from bootstrap resample split
@@ -215,15 +143,67 @@ predict_single_boot <- function(workflow,
       boot_splits$splits[[index]]
     )
 
+  # get oob sample
+  boot_oob <-
+    rsample::testing(
+      boot_splits$splits[[index]]
+    )
+
   # fit workflow to training data
   model <- generics::fit(workflow, boot_train)
+
+  # get predicted var name
+  pred_name <- dplyr::filter(workflow$pre$actions$recipe$recipe$var_info, role == "outcome")
+  pred_name <- dplyr::pull(pred_name, variable)
+
+  # get training residuals
+  preds_train <- dplyr::pull(stats::predict(model, boot_train), .pred)
+  actuals_train <- dplyr::pull(boot_train, rlang::sym(pred_name))
+  resids_train <- actuals_train - preds_train
+  resids_train <- resids_train - mean(resids_train)
+
+  # get oob residuals
+  preds_oob <- dplyr::pull(stats::predict(model, boot_oob), .pred)
+  actuals_oob <- dplyr::pull(boot_oob, rlang::sym(pred_name))
+  resids_oob <- actuals_oob - preds_oob
+  resids_oob <- resids_oob - mean(resids_oob)
+
+  # calculate no-information error rate (rmse_ni) with RMSE as loss function
+  combos <- tidyr::crossing(actuals_train, preds_train)
+  rmse_ni <- Metrics::rmse(combos$actuals_train, combos$preds_train)
+
+  # calculate overfit rate
+  rmse_oob <- Metrics::rmse(actuals_oob, preds_oob)
+  rmse_train <- Metrics::rmse(actuals_train, preds_train)
+  overfit <- (rmse_oob - rmse_train)/(rmse_ni - rmse_train)
+
+  # calculate weight (if overfit = 0, weight = .632 & residual used will just be .632)
+  # uses the actual proportion of distinct training/oob samples, rather than average of 0.632/0.368
+  prop_368 <- nrow(boot_oob)/nrow(boot_train)
+  prop_632 <- 1 - prop_368
+  weight <- prop_632/(1 - (prop_368 * overfit))
+
+  # determine residual std.dev based on weight
+  sd_oob <- stats::sd(resids_oob)
+  sd_train <- stats::sd(resids_train)
+  sd_resid <- weight * sd_oob + (1 - weight) * sd_train
 
   # predict given model and new data
   preds <- stats::predict(model, new_data)
 
+  # add residuals to fit
+  preds <- tibble::add_column(preds, resid_add = stats::rnorm(nrow(new_data), 0, sd_resid))
+  preds <- dplyr::mutate(preds, .pred = .pred + resid_add)
+  preds <- preds[, 1]
+
   # rename .pred col based on index number
   preds <- dplyr::rename(preds, !!rlang::sym(paste0(".pred_", index)) := .pred)
+
+  # print progress when verbose is set to TRUE
+  verbose_print(verbose, index, nrow(boot_splits))
 
   return(preds)
 
 }
+
+