+++ R/vi_boots.R
@@ -13,17 +13,20 @@
 #'  nested list of variable importances for each variable. The shape of the list
 #'  may vary by model type. For example, linear models return two nested columns:
 #'  the absolute value of each variable's importance and the sign (POS/NEG),
-#'  whereas XGboost models return a single nested column of variable importance.
+#'  whereas tree-based models return a single nested column of variable importance.
 #'  Similarly, the number of nested rows may vary by model type as some models
 #'  may not utilize every possible predictor.
 #'
 #' @param workflow An un-fitted workflow object.
 #' @param training_data A tibble or dataframe of data to be resampled and used for training.
 #' @param n An integer for the number of bootstrap resampled models that will be created.
+#' @param verbose A logical. Defaults to `FALSE`. If set to `TRUE`, prints progress
+#'   of training to console.
 #' @param ... Additional params passed to `rsample::bootstraps()`.
 #'
 #' @export
 #'
+#' @importFrom rlang warn
 #' @importFrom rsample bootstraps
 #' @importFrom purrr map_dfr
 #' @importFrom dplyr rename_with
@@ -45,14 +48,24 @@
 #'   vi_boots(n = 125, training_data = mtcars)
 #' }
 vi_boots <- function(workflow,
-                     n = 100,
+                     n = 2000,
                      training_data,
+                     verbose = FALSE,
                      ...) {
 
   # check arguments
   assert_workflow(workflow)
   assert_n(n)
-  assert_pred_data(workflow, training_data)
+  assert_pred_data(workflow, training_data, "training")
+
+  # warn if low n
+  if (n < 2000) {
+
+    rlang::warn(
+      paste0("At least 2000 resamples recommended for stable results.")
+    )
+
+  }
 
   # create resamples from training set
   training_boots <-
@@ -70,15 +83,17 @@ vi_boots <- function(workflow,
       ~vi_single_boot(
         workflow = workflow,
         boot_splits = training_boots,
+        verbose = verbose,
         index = .x
       )
     )
 
   # rename cols
   bootstrap_vi <- dplyr::rename_with(bootstrap_vi, tolower)
+  bootstrap_vi <- dplyr::rename(bootstrap_vi, model.importance = importance)
 
   # return a nested tibble
-  bootstrap_vi <- tidyr::nest(bootstrap_vi, importance = -variable)
+  bootstrap_vi <- tidyr::nest(bootstrap_vi, .importances = -variable)
 
   return(bootstrap_vi)
 
@@ -88,17 +103,19 @@ vi_boots <- function(workflow,
 
 #' Fit a model and get the variable importance based on a single bootstrap resample
 #'
-#' @param workflow An un-fitted workflow object.
-#' @param boot_splits A bootstrap split object created by `rsample::bootstraps()`.
-#' @param index Index of `boot_splits` to use for training
+#' @param workflow passed from `vi_boots()`
+#' @param boot_splits passed from `vi_boots()`
+#' @param verbose passed from `vi_boots()`
+#' @param index passed from `vi_boots()`
 #'
 #' @importFrom rsample training
 #' @importFrom generics fit
 #' @importFrom vip vi
-#' @importFrom workflows pull_workflow_fit
+#' @importFrom workflows extract_fit_engine
 #'
 vi_single_boot <- function(workflow,
                            boot_splits,
+                           verbose,
                            index) {
 
   # get training data from bootstrap resample split
@@ -112,6 +129,16 @@ vi_single_boot <- function(workflow,
 
   # get the variable importance from the model
   vi_boot <- vip::vi(workflows::extract_fit_engine(model))
+
+  # add model name
+  vi_boot <- dplyr::mutate(vi_boot, model = paste0(".importance_", index))
+
+  # print progress when verbose is set to TRUE
+  verbose_print(
+    verbose = verbose,
+    index = index,
+    total = nrow(boot_splits)
+  )
 
   return(vi_boot)