+++ NEWS.md
@@ -1,3 +1,20 @@
+# workboots 0.1.1
+
+### Function updates
+
+* Updates to `predict_boots()`
+  + updated function to generate prediction interval (previously was generating a prediction's confidence interval)
+  + updated default setting to assume residuals are normally distributed (can draw from the actual residual distribution by setting `normal_resid = TRUE`)
+  + updated default number of resamples
+  + updated function to draw residuals based on the [632+ rule](https://stats.stackexchange.com/questions/96739/what-is-the-632-rule-in-bootstrapping) (previously was using training residuals)
+* updated default number of resamples in `vi_boots()`
+* added param `verbose` to both `predict_boots()` and `vi_boots()` to display progress in the console. 
+* added new function `summarise_importance()`, as well alias `summarize_*` for `summarise_*`
+
+### Bug fixes
+
+* Fixed bug in `assert_pred_data()` that caused some workflows to be rejected by `predict_boots()`
+
 # workboots 0.1.0
 
 * Initial release